update for new structure

+++ Deliciousstumble/src/main/java/app/service/PostsService.java
@@ -4,9 +4,13 @@ import java.util.List;
 
 import app.domain.Post;
 
+/**
+ * @author daliborn
+ *	class used for interaction with Post objects stored locally
+ */
 public interface PostsService {
 	public void createPosts(List<Post> postsList);
 	public List<Post> getAllPosts();
 	public void updatePost(Post post);
-	public Post getPostById(Integer postId);
+	public Post getPostById(Long postId);
 }
